# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmotionalWeather = typeof EmotionalWeather[keyof typeof EmotionalWeather];


export const EmotionalWeather = {
  foggy_morning: 'foggy_morning',
  thunderstorm: 'thunderstorm',
  persistent_drizzle: 'persistent_drizzle',
  clear_skies: 'clear_skies',
  autumn_drift: 'autumn_drift',
  cold_snap: 'cold_snap',
  golden_hour: 'golden_hour',
} as const;

export type GoblinState = typeof GoblinState[keyof typeof GoblinState];


export const GoblinState = {
  emotionally_crunchy: 'emotionally_crunchy',
  unstable_moist_goblin: 'unstable_moist_goblin',
  roadkill_crunch: 'roadkill_crunch',
  vaguely_employed: 'vaguely_employed',
  suspiciously_functional: 'suspiciously_functional',
  haunted_but_hydrated: 'haunted_but_hydrated',
  tiny_forest_menace: 'tiny_forest_menace',
} as const;

export type QuestCategory = typeof QuestCategory[keyof typeof QuestCategory];


export const QuestCategory = {
  daily_rituals: 'daily_rituals',
  self_care: 'self_care',
  goblin_survival: 'goblin_survival',
  work: 'work',
  creative: 'creative',
  cave_restoration: 'cave_restoration',
  bookish: 'bookish',
} as const;

export interface DailyState {
  id: number;
  date: string;
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface DailyStateInput {
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  note?: string;
}

export interface Quest {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: QuestCategory;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  xpReward: number;
  createdAt: string;
}

export interface QuestInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  category: QuestCategory;
  xpReward?: number;
}

export interface QuestUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: QuestCategory;
  completed?: boolean;
}

export interface Character {
  id: number;
  name: string;
  title: string;
  xp: number;
  level: number;
  xpToNextLevel: number;
  questsCompleted: number;
  createdAt: string;
}

export interface QuestCompletion {
  quest: Quest;
  xpAwarded: number;
  character: Character;
}

export type QuestRecommendationsGuidance = {
  description: string;
  translation: string;
  whatYouNeed: string;
  tinyThing: string;
  successToday: string;
  recommendedCategories: QuestCategory[];
  avoid: string;
};

export interface QuestRecommendations {
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  guidance: QuestRecommendationsGuidance;
  tinyThing: string;
  successDefinition: string;
  recommendedCategories: QuestCategory[];
  quests: Quest[];
}

export interface JournalEntry {
  id: number;
  content: string;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  emotionalWeather?: string | null;
  /** @nullable */
  goblinState?: string | null;
  createdAt: string;
}

export interface JournalEntryInput {
  /** @minLength 1 */
  content: string;
  mood?: string;
  emotionalWeather?: string;
  goblinState?: string;
}

export interface CharacterUpdate {
  /** @minLength 1 */
  name?: string;
  title?: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  quest_completed: 'quest_completed',
  journal_entry: 'journal_entry',
  state_set: 'state_set',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  xpAwarded?: number | null;
  timestamp: string;
}

export interface ActivityFeed {
  items: ActivityItem[];
}

export interface GoblinStats {
  totalXp: number;
  level: number;
  questsCompletedToday: number;
  questsCompletedTotal: number;
  journalEntriesTotal: number;
  currentStreak: number;
  longestStreak: number;
}

export type ListQuestsParams = {
category?: string;
completed?: boolean;
};

export type ListJournalEntriesParams = {
limit?: number;
};