# lib/api-zod/src/generated/types/activityFeed.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface ActivityFeed {
  items: ActivityItem[];
}