# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  xpAwarded?: number | null;
  timestamp: string;
}