# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  title: string;
  xp: number;
  level: number;
  xpToNextLevel: number;
  questsCompleted: number;
  createdAt: string;
}