# lib/api-zod/src/generated/types/characterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface CharacterUpdate {
  /** @minLength 1 */
  name?: string;
  title?: string;
}