# lib/api-zod/src/generated/types/dailyState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionalWeather } from './emotionalWeather';
import type { GoblinState } from './goblinState';

export interface DailyState {
  id: number;
  date: string;
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}