# lib/api-zod/src/generated/types/dailyStateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionalWeather } from './emotionalWeather';
import type { GoblinState } from './goblinState';

export interface DailyStateInput {
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  note?: string;
}