# lib/api-zod/src/generated/types/goblinStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface GoblinStats {
  totalXp: number;
  level: number;
  questsCompletedToday: number;
  questsCompletedTotal: number;
  journalEntriesTotal: number;
  currentStreak: number;
  longestStreak: number;
}