# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  content: string;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  emotionalWeather?: string | null;
  /** @nullable */
  goblinState?: string | null;
  createdAt: string;
}