# lib/api-zod/src/generated/types/quest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestCategory } from './questCategory';

export interface Quest {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: QuestCategory;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  xpReward: number;
  createdAt: string;
}