# lib/api-zod/src/generated/types/questCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export type QuestCategory = typeof QuestCategory[keyof typeof QuestCategory];


export const QuestCategory = {
  daily_rituals: 'daily_rituals',
  self_care: 'self_care',
  goblin_survival: 'goblin_survival',
  work: 'work',
  creative: 'creative',
  cave_restoration: 'cave_restoration',
  bookish: 'bookish',
} as const;