# lib/api-zod/src/generated/types/questInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestCategory } from './questCategory';

export interface QuestInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  category: QuestCategory;
  xpReward?: number;
}