# lib/api-zod/src/generated/types/questRecommendations.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Goblin Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionalWeather } from './emotionalWeather';
import type { GoblinState } from './goblinState';
import type { Quest } from './quest';
import type { QuestCategory } from './questCategory';
import type { QuestRecommendationsGuidance } from './questRecommendationsGuidance';

export interface QuestRecommendations {
  emotionalWeather: EmotionalWeather;
  goblinState: GoblinState;
  guidance: QuestRecommendationsGuidance;
  tinyThing: string;
  successDefinition: string;
  recommendedCategories: QuestCategory[];
  quests: Quest[];
}